switch from alpine to debian based images for ruby-2.6-actionloop
The actionloop goproxy executable fails to run in the ruby-2.6-alpine
image, but works just fine in the debian based ruby-2.6 image.
Since action-ruby-v2.5 is based on ruby-2.5 (and not ruby-2.5-alpine)
it seems better to use the larger image anyways, so not worth
debugging why the actionloop proxy isn't running on alpine.